all interger converted to int

+++ src/env.c
@@ -8,29 +8,29 @@
 #include "gc.h"
 
 struct Var* vars;
-int64_t var_reserved_size = 256;
-int64_t var_size = 0;
+int var_reserved_size = 256;
+int var_size = 0;
 
-int64_t find_call_level = 0, resist_call_level = 0, nest_level = 0;
+int find_call_level = 0, resist_call_level = 0, nest_level = 0;
 
-int64_t* callstack;
-int64_t callstack_reserved_size = 256;
-int64_t callstack_size = 0;
+int* callstack;
+int callstack_reserved_size = 256;
+int callstack_size = 0;
 
 bool is_head = true;
 // internal functions
 // 変数の参照の実装
 // クロージャなので定義元の変数も参照できる。
 //
-int64_t find_idx (char* key) {
+int find_idx (char* key) {
 	//call_levelが同じ間だけルックアップをする
-	for (int64_t i = var_size - 1; i >= 0 && vars[i].call == find_call_level; --i) {
+	for (int i = var_size - 1; i >= 0 && vars[i].call == find_call_level; --i) {
 		if (strcmp(key, vars[i].key) == 0)
 			return i;
 	}
 	//funに登録されている定義場所情報から定義元変数を参照する
 	if (callstack_size > 0) {
-		for (int64_t i = callstack[callstack_size-1]; i >= 0; --i) {
+		for (int i = callstack[callstack_size-1]; i >= 0; --i) {
 			if (strcmp(key, vars[i].key) == 0)
 				return i;
 		}
@@ -41,7 +41,7 @@ int64_t find_idx (char* key) {
 // exernal functions
 void env_init() {
 	INIT(struct Var, vars, var_reserved_size);
-	INIT(int64_t, callstack, callstack_reserved_size);
+	INIT(int, callstack, callstack_reserved_size);
 
 	resist("+", alloc_bfun(Add));
 	resist("-", alloc_bfun(Sub));
@@ -69,7 +69,7 @@ void env_init() {
 }
 
 void env_quit() {
-	for (int64_t i = 0; i < var_size; ++i) {
+	for (int i = 0; i < var_size; ++i) {
 		free(vars[i].key);
 		gc_free(vars[i].node);
 	}
@@ -83,7 +83,7 @@ void into_scope() {
 }
 
 void exit_scope() {
-	for (int64_t i = var_size - 1; i >= 0; --i) {
+	for (int i = var_size - 1; i >= 0; --i) {
 		//変数はGCの管理対象外なので参照しなくなった時点で消す
 		free(vars[i].key);
 		--var_size;
@@ -98,20 +98,20 @@ void start_resist_real_arg() {
 	++resist_call_level;
 }
 
-uint32_t resist_real_arg(char* key, struct Node* node) {
+int resist_real_arg(char* key, struct Node* node) {
 	struct Var var = {is_head, deep_copy(key), node, resist_call_level, 0};
 	APPEND(struct Var, vars, var_reserved_size, var_size, var);
 	return var_size - 1;
 }
 
-bool enter_func(uint32_t pos) {
+bool enter_func(int pos) {
 	find_call_level = resist_call_level;
-	APPEND(int64_t, callstack, callstack_reserved_size, callstack_size, pos);
+	APPEND(int, callstack, callstack_reserved_size, callstack_size, pos);
 	return true;
 }
 
 void exit_func() {
-	for(int64_t i = var_size; i >= 0; --i) {
+	for(int i = var_size; i >= 0; --i) {
 		if (vars[i].call >= find_call_level) {
 			//変数名はGCの管理対象外なので参照しなくなった時点で消す
 			free(vars[i].key);
@@ -126,13 +126,13 @@ void exit_func() {
 }
 
 struct Node* find(char* key) {
-	int64_t idx;
+	int idx;
 	if ((idx = find_idx(key)) < 0)
 		return NULL;
 	return vars[idx].node;
 }
 
-uint32_t resist(char* key, struct Node* node) {
+int resist(char* key, struct Node* node) {
 	struct Var var = {is_head, deep_copy(key), node, resist_call_level, nest_level};
 	APPEND(struct Var, vars, var_reserved_size, var_size, var);
 	is_head = false;
@@ -140,24 +140,24 @@ uint32_t resist(char* key, struct Node* node) {
 }
 
 
-uint32_t current_fptr() {
+int current_fptr() {
 	return var_size;
 }
 
 void env_dump() {
 	printf("------------ env dump -------------\n");
-	printf("rcall %ld fcall %ld nest %ld\n", resist_call_level, find_call_level, nest_level);
+	printf("rcall %d fcall %d nest %d\n", resist_call_level, find_call_level, nest_level);
 	printf("----------- vars dump -------------\n");
 	printf("                name | call | nest\n");
-	for (int64_t i = 0; i < var_size; ++i) {
+	for (int i = 0; i < var_size; ++i) {
 		printf("%20s | %4d | %3d\n", vars[i].key, vars[i].call, vars[i].nest);
 	}
 	printf("----------- stack dump -------------\n");
 	if (callstack_size > 0) {
-		for (int64_t i = 0; i < callstack_size - 1; ++i) {
-			printf("%ld ->", callstack[i]);
+		for (int i = 0; i < callstack_size - 1; ++i) {
+			printf("%d ->", callstack[i]);
 		}
-		printf("%ld\n", callstack[callstack_size]);
+		printf("%d\n", callstack[callstack_size]);
 	}
 }
